Revert "Optimize drawing of small bitmaps with alpha in wxMSW"
This reverts commit d8d1226297d0bc34a32e8c5c8833a7b8bca05c8e because it
results in crashes due to writing out of bounds of the bitmap: nothing
guarantees that the entire (x, y, dstWidth, dstHeight) rectangle fits in the
destination bitmap and so the code of this commit could merrily overflow it
and did it as could be seen e.g. in the HTML test sample after scrolling
around a little.

See #16766.