Fix double clicking over checkboxes in generic wxDataViewCtrl
Double clicks over checkboxes (i.e. items using wxDataViewToggleRenderer) were
ignored because they were translated to wxEVT_DATAVIEW_ITEM_ACTIVATED events
and nothing else happened, even if the event was not processed at all.

Fix this by continuing to process double clicks as normal clicks if there is
no special handling for the activation event.

In practice this means that half of the clicks doesn't seem to be "lost" any
more when clicking a checkbox in a quick succession.