Fix the length returned by UTF-32 conversion for non-BMP input
Don't optimize the required length as this is a tiny gain resulting in big
problems with the strings containing surrogates for which the actual result is
shorter than the length returned, resulting in extra NUL bytes at the end of
the converted buffer.

This is similar to 420db09f88e165728b390ba5998331b6670aa028 (see #16298) but
for UTF-32 and not UTF-16.

Closes #17070.