Ensure that wxComboCtrl is aligned with other controls under OS X
Positioning wxComboCtrl and wxTextCtrl or wxComboBox in the same column should
align their left borders, but it didn't because of an extra offset for the
focus ring used for wxComboCtrl only.

We need to either apply this offset to all controls or none of them, so remove
it from this one.

Closes #15017.