Fix retrieving wxMask from wxBitmap in wxMSW
In wxMask::GetBitmap() we need to create a copy of the mask bitmap and assign
it to the resulting wxBitmap. We cannot simply assign existing mask bitmap to
the resulting wxBitmap because when wxMask::GetBitmap would be called more
then once then many independent wxBitmaps would own the same bitmap.

Closes https://github.com/wxWidgets/wxWidgets/pull/230

Closes #17395.