Use information icon for notifications created using default ctor
The refactoring in https://github.com/wxWidgets/wxWidgets/pull/92 changed the
appearance of wxNotificationMessages created using default ctor under MSW:
they now had no icon instead of using the default one.

Restore the old behaviour for compatibility and also because it's compatible
with the notifications created using non-default ctor but without providing
an explicit flags parameter value.

It is a bit annoying to have to use wxICON_INFORMATION explicitly both here
and as the default parameter value in include/wx/notifmsg.h, but not annoying
enough to do anything more complicated to solve it.