Disable 2 warnings occurring in release builds
Because of using a jump in an asm block there is a warning regarding
disabling global optimisations (for that function only) and as a result of
that another warning about being unable to check for buffer overruns. To
work around the warnings temporarily disable both (C4740 and C4748).

Regression since 3f5e1ada2e0ca283d5fb0672f190810ddc94741e.