Invalidate wxOSX wxListBox best size after adding items to it
Make wxListBox behave the same as in the other ports and invalidate its best
size after inserting items into it.

In the future it would be nice to call this from the base wxWindowWithItems
class itself, rather than doing it in port-specific DoInsertItems() for all
ports, but for now this will do.

Closes #17606.