Fix GetInstallPrefix availability check
Add wxHAS_STDPATHS_INSTALL_PREFIX to make checking for
wxStandardPaths::GetInstallPrefix()'s availability simpler, as the
condition under which it is compiled is nontrivial.

Fixes compilation after 86763171ac54e0e1b46567c780e46f255295c05c on
Cygwin, which is a UNIX, but its wxStandardPaths implementation is a
Windows one.