Limit the use of "webkit2_ext" to autoconf only in bakefiles
This is used by WebKit2 backend under Unix with configure, but is not needed
under the other platforms and prevented bakefile_gen from generating all
output files due to the of undefined USE_WEBVIEW_WEBKIT2 variable.

We could define it for formats other than autoconf, but this doesn't seem to
be useful currently, so just avoid referencing it instead.

Closes #17860.