Fix bug due to undefined g_wxAnyValueTypeGlobals initialization order
This fixes the changes of commit 6cd8d13776837c76dac1063f3e261232beef178a
which didn't work if any of the globals using wxPreRegisterAnyToVariant() were
instantiated before g_wxAnyValueTypeGlobals itself.

Wrap the global inside a function to ensure that we initialized it before it
is used by wxPreRegisterAnyToVariant() and not after.